+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,53 @@
 # sv
 
+## 0.6.25
+### Patch Changes
+
+
+- chore: detect package manager asynchronously ([#465](https://github.com/sveltejs/cli/pull/465))
+
+
+- chore: add keys to `{#each}` blocks ([#466](https://github.com/sveltejs/cli/pull/466))
+
+
+- fix: pass `schema` to `drizzle` client for better type generation ([#459](https://github.com/sveltejs/cli/pull/459))
+
+
+- fix: addons executed in the wrong order in certain circumstances ([#462](https://github.com/sveltejs/cli/pull/462))
+
+## 0.6.24
+### Patch Changes
+
+
+- fix: use 'prettier' instead of ['flat/prettier'] ([#467](https://github.com/sveltejs/cli/pull/467))
+
+
+- fix: properly add tailwind plugins on subsequent add-on executions ([#456](https://github.com/sveltejs/cli/pull/456))
+
+## 0.6.23
+### Patch Changes
+
+
+- fix: use `eslint-plugin-svelte` v3 ([#455](https://github.com/sveltejs/cli/pull/455))
+
+## 0.6.22
+### Patch Changes
+
+
+- fix: make `drizzle` next steps more precise ([#447](https://github.com/sveltejs/cli/pull/447))
+
+
+- chore: update addon dependencies ([#450](https://github.com/sveltejs/cli/pull/450))
+
+
+- fix: allow selecting adapter auto ([#446](https://github.com/sveltejs/cli/pull/446))
+
+
+- feat: update `eslint-plugin-svelte` v3 ([#453](https://github.com/sveltejs/cli/pull/453))
+
+
+- fix: re-add `tailwindcss` plugins ([#448](https://github.com/sveltejs/cli/pull/448))
+
 ## 0.6.21
 ### Patch Changes